backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
tavily-python==0.3.3
sentence-transformers==2.7.0
scikit-learn==1.4.2
nltk==3.8.1
numpy==1.26.4
httpx==0.27.0
pydantic==2.7.1
python-multipart==0.0.9
qdrant-client==1.9.0
groq==0.8.0
openai==1.30.1
pypdf==4.2.0
